Move scheduler util to own file
